+++ src/onegov.form/onegov/form/tests/test_fields.py
@@ -5,7 +5,7 @@ from cgi import FieldStorage
 from gzip import GzipFile
 from io import BytesIO
 from onegov.form import Form
-from onegov.form.fields import UploadField
+from onegov.form.fields import CoordinateField, UploadField
 
 
 def create_file(mimetype, filename, content):
@@ -35,3 +35,30 @@ def test_upload_file():
             return f.read()
 
     assert decompress(base64.b64decode(data['data'])) == b'foobar'
+
+
+def test_coordinate_field():
+    field = CoordinateField().bind(Form(), 'coordinate')
+    template = (
+        '<input class="coordinate" id="coordinate" name="coordinate" '
+        'type="text" value="{}">'
+    )
+
+    assert not field.data
+
+    assert field.data.lat is None
+    assert field.data.lon is None
+    assert field() == template.format('')
+
+    field.process_formdata(['47.05183585/8.30576869173879'])
+
+    assert field.data.lat == 47.05183585
+    assert field.data.lon == 8.30576869173879
+
+    field.data.lat = '47.05183585'
+    assert field.data.lat == 47.05183585
+
+    field.data.lon = '8.30576869173879'
+    assert field.data.lon == 8.30576869173879
+
+    assert field() == template.format('47.05183585/8.30576869173879')